abstract into ui-app, separate wrappers, NominationsList
