+++ packages/app-staking/src/library/NominationList/DetailedItem.tsx
@@ -1,22 +1,26 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import classNames from 'classnames'
 import { getStakingChainData } from 'consts/util'
-import { useList } from 'contexts/List'
 import { useValidators } from 'contexts/Validators/ValidatorEntries'
 import { useNetwork } from 'hooks/useNetwork'
-import { HistoricalEraPoints } from 'library/List/EraPointsGraph/HistoricalEraPoints'
 import { getIdentityDisplay } from 'library/List/Utils'
 import { CopyAddress } from 'library/ListItem/Buttons/CopyAddress'
+import { FavoriteValidator } from 'library/ListItem/Buttons/FavoriteValidator'
 import { Metrics } from 'library/ListItem/Buttons/Metrics'
-import { Remove } from 'library/ListItem/Buttons/Remove'
-import { ShareLink } from 'library/ListItem/Buttons/ShareLink'
+import { Identity } from 'library/ListItem/Labels/Identity'
+import { useNominationStatusData } from 'library/ListItem/Labels/NominationStatus'
+import { RetainmentStats } from 'library/ValidatorList/RetainmentStats'
+import { RowActionsMenu } from 'library/ValidatorList/RowActionsMenu'
+import { useRetainmentStatsData } from 'library/ValidatorList/useRetainmentStatsData'
+import { useValidatorSelfStake } from 'library/ValidatorList/useValidatorSelfStake'
+import { ValidatorBar } from 'library/ValidatorList/ValidatorBar'
+import { ValidatorSummary } from 'library/ValidatorList/ValidatorSummary'
 import { useTranslation } from 'react-i18next'
-import type { Validator } from 'types'
 import {
 	BlockedBadge,
 	CardTop,
+	DetailLoader,
 	HeaderActions,
 	HeaderIconAction,
 	HeaderIdentity,
@@ -26,58 +30,46 @@ import {
 	PerformanceRow,
 	SectionHeader,
 } from 'ui-app/ListItem'
-import { FavoriteValidator } from '../ListItem/Buttons/FavoriteValidator'
-import { Select } from '../ListItem/Buttons/Select'
-import { Identity } from '../ListItem/Labels/Identity'
-import { DetailedItemPreloader } from './DetailedItemPreloader'
-import { RetainmentStats } from './RetainmentStats'
-import { RowActionsMenu } from './RowActionsMenu'
+import { HistoricalEraPoints } from '../List/EraPointsGraph/HistoricalEraPoints'
 import type { ItemProps } from './types'
-import { useRetainmentStatsData } from './useRetainmentStatsData'
-import { useValidatorSelfStake } from './useValidatorSelfStake'
-import { ValidatorBar } from './ValidatorBar'
-import { ValidatorSummary } from './ValidatorSummary'
 
 export const DetailedItem = ({
 	validator,
+	nominator,
 	toggleFavorites,
+	bondFor,
 	displayFor,
-	eraPoints,
-	onRemove,
-	rate,
 	format,
+	nominationStatus = 'waiting',
+	eraPoints,
+	rate,
 	retainment,
-	isPreloading,
+	isPreloading = false,
 }: ItemProps) => {
 	const { t } = useTranslation('app')
 	const { network } = useNetwork()
-	const { selectable, selected } = useList()
 	const { validatorIdentities, validatorSupers } = useValidators()
 	const { address, prefs, validatorStatus } = validator
 	const commission = prefs?.commission ?? null
 	const { unit, units } = getStakingChainData(network)
 	const { selfStake, selfStakeMax } = useValidatorSelfStake(address, units)
+	const {
+		label: statusLabel,
+		stakedAmount: backingStake,
+		syncing: backingStakePreloading,
+	} = useNominationStatusData({
+		address,
+		bondFor,
+		nominator,
+		status: nominationStatus,
+	})
 	const retainmentStats = useRetainmentStatsData({
 		period: retainment?.months[0],
 		selfStakeMax,
 		unit,
 		units,
 	})
-
-	if (isPreloading) {
-		return <DetailedItemPreloader format={format} />
-	}
-
-	const isSelected = !!selected.filter(
-		(item) => (item as Validator).address === validator.address,
-	).length
-
-	const innerClasses = classNames('inner', {
-		[displayFor]: true,
-		selected: isSelected,
-	})
-
-	// Rate after commission
+	const outline = displayFor === 'canvas'
 	const rateAfterCommission =
 		typeof rate === 'number' &&
 		Number.isFinite(rate) &&
@@ -90,36 +82,6 @@ export const DetailedItem = ({
 		validatorSupers[address],
 	).node
 
-	const cardActions = (
-		<HeaderActions>
-			<HeaderIconAction>
-				<CopyAddress address={address} />
-			</HeaderIconAction>
-			<HeaderIconAction>
-				<ShareLink paramKey="v" paramValue={address} />
-			</HeaderIconAction>
-			{toggleFavorites && (
-				<HeaderIconAction>
-					<FavoriteValidator address={address} />
-				</HeaderIconAction>
-			)}
-			{typeof onRemove === 'function' && (
-				<HeaderIconAction>
-					<Remove
-						address={address}
-						onRemove={() => onRemove({ selected: [validator] })}
-						displayFor={displayFor}
-					/>
-				</HeaderIconAction>
-			)}
-			{displayFor === 'default' && (
-				<HeaderMetricsAction>
-					<Metrics address={address} display={validatorDisplay} />
-				</HeaderMetricsAction>
-			)}
-		</HeaderActions>
-	)
-
 	if (format === 'row') {
 		return (
 			<ValidatorBar
@@ -127,22 +89,22 @@ export const DetailedItem = ({
 					<RowActionsMenu
 						address={address}
 						display={validatorDisplay}
-						onRemove={
-							typeof onRemove === 'function'
-								? () => onRemove({ selected: [validator] })
-								: undefined
-						}
 						showFavorite={toggleFavorites === true}
-						showMetrics={displayFor === 'default'}
+						showMetrics={displayFor !== 'canvas'}
 					/>
 				}
-				innerClasses={innerClasses}
 				displayFor={displayFor}
 				eraPoints={eraPoints}
+				innerClasses={`inner ${displayFor}`}
+				isPreloading={isPreloading}
+				isStatusValuePreloading={backingStakePreloading}
 				rate={rateAfterCommission}
 				retainmentStats={retainmentStats}
 				selfStake={selfStake}
 				selfStakeMax={selfStakeMax}
+				statusActive={nominationStatus === 'active'}
+				statusLabel={statusLabel}
+				statusValue={backingStake}
 				unit={unit}
 				validator={validator}
 			/>
@@ -151,41 +113,80 @@ export const DetailedItem = ({
 
 	return (
 		<ItemWrapper>
-			<div className={innerClasses}>
+			<div className={`inner ${displayFor}`}>
 				<CardTop className="card-top">
 					<div className="row top">
-						{selectable && <Select item={validator} />}
 						<HeaderIdentity>
 							<Identity address={address} />
 							{prefs?.blocked === true && (
 								<BlockedBadge>{t('blocked')}</BlockedBadge>
 							)}
 						</HeaderIdentity>
-						{cardActions}
+						<HeaderActions>
+							<HeaderIconAction>
+								<CopyAddress address={address} />
+							</HeaderIconAction>
+							{toggleFavorites && (
+								<HeaderIconAction>
+									<FavoriteValidator address={address} outline={outline} />
+								</HeaderIconAction>
+							)}
+							{displayFor !== 'canvas' && (
+								<HeaderMetricsAction>
+									<Metrics
+										address={address}
+										display={validatorDisplay}
+										outline={outline}
+									/>
+								</HeaderMetricsAction>
+							)}
+						</HeaderActions>
 					</div>
 					<ValidatorSummary
 						address={address}
+						ariaLabel={t('nominationSummary', {
+							defaultValue: 'Nomination summary',
+						})}
+						isRatePreloading={isPreloading}
+						isStatusValuePreloading={backingStakePreloading}
 						rate={rateAfterCommission}
 						selfStake={selfStake}
 						selfStakeMax={selfStakeMax}
 						status={validatorStatus}
+						statusActive={nominationStatus === 'active'}
+						statusLabel={statusLabel}
+						statusValue={backingStake}
 						unit={unit}
 					/>
-					<PerformanceRow className="row performance">
+					<PerformanceRow className="row performance" aria-busy={isPreloading}>
 						<SectionHeader>
 							<strong>{t('performance')}</strong>
 						</SectionHeader>
 						<PerformanceGraph>
-							<HistoricalEraPoints
-								address={address}
-								displayFor={displayFor}
-								eraPoints={eraPoints}
-								stretch
-							/>
+							{isPreloading ? (
+								<div>
+									<DetailLoader
+										borderRadius="0.45rem"
+										height="100%"
+										width="100%"
+									/>
+								</div>
+							) : (
+								<HistoricalEraPoints
+									address={address}
+									displayFor={displayFor}
+									eraPoints={eraPoints}
+									stretch
+								/>
+							)}
 						</PerformanceGraph>
 					</PerformanceRow>
 				</CardTop>
-				<RetainmentStats data={retainmentStats} unit={unit} />
+				<RetainmentStats
+					data={retainmentStats}
+					isPreloading={isPreloading}
+					unit={unit}
+				/>
 			</div>
 		</ItemWrapper>
 	)
