+++ packages/app-staking/src/library/NominationList/types.ts
@@ -1,8 +1,12 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import type { ListFormat } from 'contexts/List/types'
 import type { ValidatorListEntry } from 'contexts/Validators/types'
-import type { ValidatorEraPoints } from 'plugin-staking-api/types'
+import type {
+	ValidatorEraPoints,
+	ValidatorRetainmentResult,
+} from 'plugin-staking-api/types'
 import type {
 	BondFor,
 	DisplayFor,
@@ -27,8 +31,11 @@ export interface ItemProps {
 	nominator: MaybeAddress
 	bondFor: BondFor
 	displayFor: DisplayFor
+	format: ListFormat
 	toggleFavorites?: boolean
 	nominationStatus?: NominationStatus
 	eraPoints: ValidatorEraPoints[]
+	isPreloading?: boolean
 	rate?: number
+	retainment?: ValidatorRetainmentResult | null
 }
