+++ packages/app-staking/src/library/ValidatorList/RetainmentStats.tsx
@@ -1,12 +1,8 @@
 // Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { RetainmentStatValue } from './RetainmentStatValue'
-import type {
-	RetainmentStatData,
-	RetainmentStatsData,
-} from './useRetainmentStatsData'
 import {
+	DetailLoader,
 	FlowLabel,
 	FlowMetric,
 	FlowValue,
@@ -14,14 +10,41 @@ import {
 	RetainmentBody,
 	RetainmentRow,
 	SectionHeader,
-} from './Wrappers'
+} from 'ui-app/ListItem'
+import { RetainmentStatValue } from './RetainmentStatValue'
+import type {
+	RetainmentStatData,
+	RetainmentStatsData,
+} from './useRetainmentStatsData'
 
 interface RetainmentStatsProps {
 	data: RetainmentStatsData
+	isPreloading?: boolean
 	unit: string
 }
 
-const RateStat = ({ stat }: { stat: RetainmentStatData }) => {
+const RetainmentValue = ({
+	isPreloading,
+	stat,
+	unit,
+}: {
+	isPreloading: boolean
+	stat: RetainmentStatData
+	unit?: string
+}) =>
+	isPreloading ? (
+		<DetailLoader />
+	) : (
+		<RetainmentStatValue stat={stat} unit={unit} />
+	)
+
+const RateStat = ({
+	isPreloading,
+	stat,
+}: {
+	isPreloading: boolean
+	stat: RetainmentStatData
+}) => {
 	return (
 		<FlowMetric>
 			<FlowLabel title={stat.label}>{stat.label}</FlowLabel>
@@ -34,16 +57,18 @@ const RateStat = ({ stat }: { stat: RetainmentStatData }) => {
 				aria-valuenow={stat.value}
 				aria-valuetext={stat.ariaValueText}
 			>
-				<RetainmentStatValue stat={stat} />
+				<RetainmentValue isPreloading={isPreloading} stat={stat} />
 			</FlowValue>
 		</FlowMetric>
 	)
 }
 
 const SignedAmountStat = ({
+	isPreloading,
 	stat,
 	unit,
 }: {
+	isPreloading: boolean
 	stat: RetainmentStatData
 	unit: string
 }) => {
@@ -51,13 +76,17 @@ const SignedAmountStat = ({
 		<FlowMetric>
 			<FlowLabel title={stat.label}>{stat.label}</FlowLabel>
 			<FlowValue $color={stat.color} aria-label={stat.ariaLabel}>
-				<RetainmentStatValue stat={stat} unit={unit} />
+				<RetainmentValue isPreloading={isPreloading} stat={stat} unit={unit} />
 			</FlowValue>
 		</FlowMetric>
 	)
 }
 
-export const RetainmentStats = ({ data, unit }: RetainmentStatsProps) => {
+export const RetainmentStats = ({
+	data,
+	isPreloading = false,
+	unit,
+}: RetainmentStatsProps) => {
 	const {
 		compoundRate,
 		month,
@@ -69,20 +98,34 @@ export const RetainmentStats = ({ data, unit }: RetainmentStatsProps) => {
 	} = data
 
 	return (
-		<RetainmentRow className="row retainment" aria-label={statsLabel}>
+		<RetainmentRow
+			className="row retainment"
+			aria-busy={isPreloading}
+			aria-label={statsLabel}
+		>
 			<SectionHeader>
 				<strong>{retainmentLabel}</strong>
-				{month && (
+				{isPreloading ? (
+					<DetailLoader height="0.85rem" width="7rem" />
+				) : month ? (
 					<MonthBadge dateTime={month.date.toISOString()}>
 						/ {month.label}
 					</MonthBadge>
-				)}
+				) : null}
 			</SectionHeader>
 			<RetainmentBody>
-				<RateStat stat={retainmentRate} />
-				<RateStat stat={compoundRate} />
-				<SignedAmountStat stat={selfStakeChange} unit={unit} />
-				<SignedAmountStat stat={netOutflow} unit={unit} />
+				<RateStat isPreloading={isPreloading} stat={retainmentRate} />
+				<RateStat isPreloading={isPreloading} stat={compoundRate} />
+				<SignedAmountStat
+					isPreloading={isPreloading}
+					stat={selfStakeChange}
+					unit={unit}
+				/>
+				<SignedAmountStat
+					isPreloading={isPreloading}
+					stat={netOutflow}
+					unit={unit}
+				/>
 			</RetainmentBody>
 		</RetainmentRow>
 	)
