+++ packages/app-staking/src/library/ValidatorList/ValidatorBar.tsx
@@ -9,14 +9,6 @@ import type { ValidatorEraPoints } from 'plugin-staking-api/types'
 import type { ReactNode } from 'react'
 import { useTranslation } from 'react-i18next'
 import type { DisplayFor } from 'types'
-import { Select } from '../ListItem/Buttons/Select'
-import { Identity } from '../ListItem/Labels/Identity'
-import { RetainmentStatValue } from './RetainmentStatValue'
-import type {
-	RetainmentStatData,
-	RetainmentStatsData,
-} from './useRetainmentStatsData'
-import { useValidatorSummaryData } from './ValidatorSummary'
 import {
 	BarIdentity,
 	BarLayout,
@@ -27,24 +19,44 @@ import {
 	BarStatValue,
 	BarWrapper,
 	BlockedBadge,
+	DetailLoader,
 	HeaderIdentity,
 	SummaryStatusDot,
-} from './Wrappers'
+} from 'ui-app/ListItem'
+import { Select } from '../ListItem/Buttons/Select'
+import { Identity } from '../ListItem/Labels/Identity'
+import { RetainmentStatValue } from './RetainmentStatValue'
+import type {
+	RetainmentStatData,
+	RetainmentStatsData,
+} from './useRetainmentStatsData'
+import { useValidatorSummaryData } from './ValidatorSummary'
 
 interface ValidatorBarProps {
 	actions: ReactNode
 	displayFor: DisplayFor
 	eraPoints: ValidatorEraPoints[]
 	innerClasses: string
+	isPreloading?: boolean
+	isStatusValuePreloading?: boolean
 	rate?: number
 	retainmentStats: RetainmentStatsData
 	selfStake?: BigNumber
 	selfStakeMax: boolean
+	statusActive?: boolean
+	statusLabel?: string
+	statusValue?: BigNumber
 	unit: string
 	validator: ValidatorListEntry
 }
 
-const BarRateStat = ({ stat }: { stat: RetainmentStatData }) => (
+const BarRateStat = ({
+	isPreloading,
+	stat,
+}: {
+	isPreloading: boolean
+	stat: RetainmentStatData
+}) => (
 	<BarStat>
 		<BarStatLabel title={stat.label}>{stat.label}</BarStatLabel>
 		<BarStatValue
@@ -56,22 +68,32 @@ const BarRateStat = ({ stat }: { stat: RetainmentStatData }) => (
 			aria-valuenow={stat.value}
 			aria-valuetext={stat.ariaValueText}
 		>
-			<RetainmentStatValue stat={stat} />
+			{isPreloading ? (
+				<DetailLoader height="1.2rem" width="4.5rem" />
+			) : (
+				<RetainmentStatValue stat={stat} />
+			)}
 		</BarStatValue>
 	</BarStat>
 )
 
 const BarSignedAmountStat = ({
+	isPreloading,
 	stat,
 	unit,
 }: {
+	isPreloading: boolean
 	stat: RetainmentStatData
 	unit: string
 }) => (
 	<BarStat>
 		<BarStatLabel title={stat.label}>{stat.label}</BarStatLabel>
 		<BarStatValue $color={stat.color} aria-label={stat.ariaLabel}>
-			<RetainmentStatValue stat={stat} unit={unit} />
+			{isPreloading ? (
+				<DetailLoader height="1.2rem" width="4.5rem" />
+			) : (
+				<RetainmentStatValue stat={stat} unit={unit} />
+			)}
 		</BarStatValue>
 	</BarStat>
 )
@@ -81,10 +103,15 @@ export const ValidatorBar = ({
 	displayFor,
 	eraPoints,
 	innerClasses,
+	isPreloading = false,
+	isStatusValuePreloading = false,
 	rate,
 	retainmentStats,
 	selfStake,
 	selfStakeMax,
+	statusActive,
+	statusLabel,
+	statusValue,
 	unit,
 	validator,
 }: ValidatorBarProps) => {
@@ -95,7 +122,7 @@ export const ValidatorBar = ({
 		quartileLabel,
 		rateLabel,
 		selfStakeLabel,
-		statusLabel,
+		statusLabel: summaryStatusLabel,
 		totalStake,
 		validatorStatus,
 	} = useValidatorSummaryData({
@@ -104,6 +131,8 @@ export const ValidatorBar = ({
 		selfStake,
 		selfStakeMax,
 		status,
+		statusLabel,
+		statusValue,
 		unit,
 	})
 	const { compoundRate, netOutflow, retainmentRate, selfStakeChange } =
@@ -125,12 +154,20 @@ export const ValidatorBar = ({
 							aria-label={t('performance')}
 							title={t('performance')}
 						>
-							<HistoricalEraPoints
-								address={address}
-								displayFor={displayFor}
-								eraPoints={eraPoints}
-								stretch
-							/>
+							{isPreloading ? (
+								<DetailLoader
+									borderRadius="0.3rem"
+									height="100%"
+									width="100%"
+								/>
+							) : (
+								<HistoricalEraPoints
+									address={address}
+									displayFor={displayFor}
+									eraPoints={eraPoints}
+									stretch
+								/>
+							)}
 						</BarPerformanceGraph>
 					</BarIdentity>
 
@@ -138,21 +175,33 @@ export const ValidatorBar = ({
 						<BarStat>
 							<BarStatLabel>
 								<SummaryStatusDot
-									$active={validatorStatus === 'active'}
+									$active={statusActive ?? validatorStatus === 'active'}
 									aria-hidden="true"
 								/>
-								<span>{statusLabel}</span>
+								<span>{summaryStatusLabel}</span>
 							</BarStatLabel>
 							<BarStatValue
 								title={totalStake ? `${totalStake} ${unit}` : undefined}
 							>
-								<span>{totalStake ?? '—'}</span>
-								{totalStake && <small>{unit}</small>}
+								{isStatusValuePreloading ? (
+									<DetailLoader height="1.2rem" width="4.5rem" />
+								) : (
+									<>
+										<span>{totalStake ?? '—'}</span>
+										{totalStake && <small>{unit}</small>}
+									</>
+								)}
 							</BarStatValue>
 						</BarStat>
 						<BarStat>
 							<BarStatLabel>APY</BarStatLabel>
-							<BarStatValue>{rateLabel}</BarStatValue>
+							<BarStatValue>
+								{isPreloading ? (
+									<DetailLoader height="1.2rem" width="4.5rem" />
+								) : (
+									rateLabel
+								)}
+							</BarStatValue>
 						</BarStat>
 						<BarStat>
 							<BarStatLabel>{t('performance')}</BarStatLabel>
@@ -169,10 +218,18 @@ export const ValidatorBar = ({
 								)}
 							</BarStatValue>
 						</BarStat>
-						<BarRateStat stat={retainmentRate} />
-						<BarRateStat stat={compoundRate} />
-						<BarSignedAmountStat stat={selfStakeChange} unit={unit} />
-						<BarSignedAmountStat stat={netOutflow} unit={unit} />
+						<BarRateStat isPreloading={isPreloading} stat={retainmentRate} />
+						<BarRateStat isPreloading={isPreloading} stat={compoundRate} />
+						<BarSignedAmountStat
+							isPreloading={isPreloading}
+							stat={selfStakeChange}
+							unit={unit}
+						/>
+						<BarSignedAmountStat
+							isPreloading={isPreloading}
+							stat={netOutflow}
+							unit={unit}
+						/>
 					</BarStats>
 
 					{actions}
