+++ packages/utils/src/index.ts
@@ -7,6 +7,7 @@ export * from './conversions'
 export * from './datetime'
 export * from './identity'
 export * from './migrations'
+export * from './numbers'
 export * from './observableStore'
 export * from './pages'
 export * from './pools'
